# swagger.yaml
openapi: 3.0.0
info:
  title: ISDServices
  description: In order to create a post on your medical reserch work, read about other medical post and comments, ISDServices provides you this placform, you must first initiate the process from the platform by registering.
  contact:
    name: hr.admin@isdservices.org
    email: integratedsoftwaredataservices@gmail.com
  version: '1.0'
servers:
  - url: https://isdservices.org
    description: This is the production server
    variables: {}
  - url: http://localhost:4400
    description: 'This is the localhost development server'
    variables: {}
paths:
...
  /user/singleUser/{id}:
    get:
      tags:
        - User
      summary: SingleUser
      description: This returns details of a single user
      operationId: SingleUser
      parameters:
        - name: id
          in: path
          description: The user ID
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: ''
          headers: {}
          content: {}
      deprecated: false
      security: []
...
  /user/register:
    post:
      tags:
        - Auth
      summary: RegisterUser
      description: This is the signup endpoint, which allows you to register a user
      operationId: RegisterUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/registeruserrequest'
                - example:
                    name: Adeyinka
                    occupation: Kayode
                    dateOfBirth: tjvpro
                    email: example@domain.com
                    password: '1234567890'
                    confirmPassword: '1234567890'
            example:
              name: Adeyinka
              occupation: Kayode
              dateOfBirth: tjvpro
              email: example@domain.com
              password: '1234567890'
              confirmPassword: '1234567890'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content: {}
      deprecated: false
      security: []
  /user/login:
    post:
      tags:
        - Auth
      summary: LoginUser
      description: You can log in either by using a registered email or a registered dateOfBirth with a password
      operationId: LoginUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/loginuserrequest'
                - example:
                    email: email@hostname.com
                    password: userpassword
            example:
              email: email@hostname.com
              password: userpassword
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content: {}
      deprecated: false
      security: []
  /user/verify/{token}:
    get:
      tags:
        - Auth
      summary: VerifyEmail
      description: This endpoint verifies an unverified email, to get the token for verification you can go through the resend verification token endpoint or upon registration of the user, after which the user would be redirected to the login page
      operationId: VerifyEmail
      parameters:
        - name: token
          in: path
          description: The user ID
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: ''
          headers: {}
          content: {}
      deprecated: false
      security: []
...
  /user/all-users:
    get:
      tags:
        - Admin
      summary: AllUsers
      description: This endpoint returns all registered users, accessed by logged-in admin only
      operationId: AllUsers
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
...
  /user/deleteUser/{userId}:
    delete:
      tags:
        - Admin
      summary: DeleteUser
      description: This endpoint deletes a registered user from the system database, accessed only by logged-in admin
      operationId: DeleteUser
      parameters:
        - name: userId
          in: path
          description: The user ID
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: ''
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
components:
  schemas:
...
    registeruserrequest:
      title: registeruserrequest
      required:
        - name
        - occupation
        - dateOfBirth
        - email
        - password
        - confirmPassword
      type: object
      properties:
        name:
          type: string
        occupation:
          type: string
        dateOfBirth:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
      example:
        name: Adeyinka
        occupation: Doctor
        dateOfBirth: 04/04/2023
        email: example@domain.com
        password: '1234567890'
        confirmPassword: '1234567890'
    loginuserrequest:
      title: loginuserrequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: example@domain.com
        password: '1234567890'
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...